Moving around config files
