+++ src/db/bookshelf.js
@@ -2,7 +2,7 @@ console.log("Knex and Bookshelf initialization, should only happen once!");
 var Knex = require('knex');
 var Bookshelf = require('bookshelf');
 var path = require('path');
-var dbConfig = require('../knex/knexfile')[process.env.NODE_ENV || "development"];
+var dbConfig = require('../db/knexfile')[process.env.NODE_ENV || "development"];
 
 var knex = new Knex(dbConfig);
 knex.config = dbConfig;
